requirements.txt
annotated-types==0.7.0
anyio==4.7.0
certifi==2024.8.30
colorama==0.4.6
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
neo4j==5.27.0
packaging==24.2
pluggy==1.5.0
pydantic==2.10.3
pydantic_core==2.27.1
pytest==8.3.4
python-dotenv==1.0.1
pytz==2024.2
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
